[Streaming][OKCoin] refactor into parent class
